Proof-of-concept: file parsing and listing
Created a proof-of-concept for the main page.
The idea is iterating on the _posts directory (and possible
sub-directories) in the server, parsing each file and visualizing a
list of available blog-posts.
This is done by processing the old format that the blog supported when
using Jekyll and the "standard" configuration for Github Pages.

Note that, in the future, we might need to break this list down in more
pages. That might be challenging if there is no way to produce multiple
pages (or sub-pages) from this initial one.