Tentative proof-of-concept for dynamic loading
In src/app/(posts)/page.tsx, the goal is being able to import all the
posts, filtering them and showing the relevant excerpts in a list.
However, as documented in the following link:

https://stackoverflow.com/a/75566483

dynamically importing with a dynamic path is not allowed.
The idea of retrieving the path of each post dynamically, and importing
them there, is not viable.

NOTE to self: I wonder if one could generate the paths at compile
time/build time and loop over them; my understanding is that this would
not work anyway.

Possible solution: read the file, and "compile it" using MDX.